move EditableCustomer from Dto to QueryResult namespace
